source/matrix/client/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.4
aiohttp-socks==0.8.4
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
matrix-nio==0.24.0
multidict==6.0.5
numpy==1.26.4
opencv-python==4.9.0.80
paho-mqtt==2.0.0
pycryptodome==3.20.0
python-dotenv==1.0.1
python-socks==2.4.4
referencing==0.34.0
rpds-py==0.18.0
unpaddedbase64==2.1.0
yarl==1.9.4